Move TGInaccessibleMessage extension to separate file
